chore : add validator helper methods in booking service chore : add helper method to fetch all tikcetType ids in the creaetBookingRequestDTO

+++ src/main/java/com/tiko/tiko/Booking/DTOs/BookingItemRequest.java
@@ -8,6 +8,6 @@ public record BookingItemRequest(
         Long eventTicketPriceId,
 
         @Positive
-        Long quantity
+        int quantity
 ) {
 }
